requirement.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cors==1.0.1
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
future==1.0.0
gevent==25.9.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.8.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pydeck==0.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
requests-file==3.0.1
rpds-py==0.27.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.50.0
tenacity==9.1.2
tldextract==5.3.0
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
zope.event==6.0
zope.interface==8.0.1
